+++ web/tailwind.config.js
@@ -2,64 +2,25 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-const colors = require("tailwindcss/colors")
 const fontCatalog = require("./ui/fonts.js")
 
-// source:
-// https://uicolors.app/edit?sv1=golden-tainoi:50-fff9eb/100-ffefc6/200-ffdd88/300-ffcb5c/400-ffae20/500-f98a07/600-dd6402/700-b74406/800-94330c/900-7a2c0d/950-461402;mandy:50-fef2f3/100-fee5e7/200-fccfd6/300-f9a8b3/400-f5778a/500-ed4e6b/600-d9254f/700-b71942/800-99183d/900-83183a/950-49081b;pale-oyster:50-fdfcfc/100-f8f7f6/200-ded9d3/300-c0b4aa/400-a7978b/500-968273/600-81695f/700-69544f/800-594946/900-4c3f3d/950-291f1f;shakespeare:50-f2f9fd/100-e5f1f9/200-c4e4f3/300-91cee8/400-43acd6/500-309bc7/600-217da8/700-1c6488/800-1b5571/900-1b475f/950-122e3f
-const palette = {
-  blue: {
-    50: "#F2FBFC",
-    100: "#E3F8FA",
-    200: "#BDEAF0",
-    300: "#97DBE6",
-    400: "#55C1D4",
-    500: "#1ba1bf",
-    600: "#178DAD",
-    700: "#0E6A8F",
-    800: "#094E73",
-    900: "#053657",
-    950: "#021F38",
-  },
-  orange: {
-    50: "#fff9eb",
-    100: "#ffefc6",
-    200: "#ffdd88",
-    300: "#ffcb5c",
-    400: "#ffae20",
-    500: "#f98a07",
-    600: "#dd6402",
-    700: "#b74406",
-    800: "#94330c",
-    900: "#7a2c0d",
-    950: "#461402",
-  },
-  red: {
-    50: "#fef4f6",
-    100: "#fce2e7",
-    200: "#f8bdc8",
-    300: "#f498a9",
-    400: "#f0738a",
-    500: "#ed4e6b",
-    600: "#e9294c",
-    700: "#d31638",
-    800: "#a0112b",
-    900: "#6d0b1d",
-    950: "#540916",
-  },
-  gray: {
-    50: "#fdfcfc",
-    100: "#f8f7f6",
-    200: "#ded9d3",
-    300: "#c0b4aa",
-    400: "#a7978b",
-    500: "#968273",
-    600: "#81695f",
-    700: "#69544f",
-    800: "#594946",
-    900: "#4c3f3d",
-    950: "#291f1f",
-  },
+const colorVar = (name) => {
+  return ({opacityValue}) => {
+    if (opacityValue) {
+      return `rgb(var(--color-${name}) / ${opacityValue})`
+    }
+    return `rgb(var(--color-${name}))`
+  }
+}
+
+const varPalette = (name) => {
+  return [50, 100, 200, 300, 400, 500, 600, 700, 800, 900, 950].reduce(
+    (acc, cur) => {
+      acc[cur] = colorVar(`${name}-${cur}`)
+      return acc
+    },
+    {},
+  )
 }
 
 // prettier-ignore
@@ -68,21 +29,51 @@ var config = {
     "src/**/*.js",
     "../assets/templates/**/*.jet.html",
   ],
+  darkMode: "class",
   theme: {
     colors: {
+      inherit: "inherit",
       transparent: "transparent",
       current: "currentColor",
-      black: colors.black,
-      white: colors.white,
-      gray: palette.gray,
-      red: palette.red,
-      green: colors.lime,
-      blue: palette.blue,
-      yellow: palette.orange,
+      black: colorVar("black"),
+      white: colorVar("white"),
+      app: {
+        bg: colorVar("app-bg"),
+        fg: colorVar("app-fg"),
+      },
+      gray: {
+        ...varPalette("gray"),
+        "light": colorVar("gray-light"),
+        "dark": colorVar("gray-dark"),
+      },
+      red: varPalette("red"),
+      green: varPalette("green"),
+      blue: varPalette("blue"),
+      yellow: varPalette("yellow"),
       primary: {
-        light: palette.blue[300],
-        DEFAULT: palette.blue[600],
-        dark: palette.blue[800],
+        ...varPalette("blue"),
+        DEFAULT: colorVar("primary"),
+        "light": colorVar("primary-light"),
+        "dark": colorVar("primary-dark"),
+      },
+
+      // button colors
+      btn: {
+        "default": colorVar("btn-default"),
+        "default-hover": colorVar("btn-default-hover"),
+        "default-text": colorVar("btn-default-text"),
+        "primary": colorVar("btn-primary"),
+        "primary-hover": colorVar("btn-primary-hover"),
+        "primary-text": colorVar("btn-primary-text"),
+        "danger": colorVar("btn-danger"),
+        "danger-hover": colorVar("btn-danger-hover"),
+        "danger-text": colorVar("btn-danger-text"),
+      },
+
+      // highlight colors
+      hl: {
+        yellow: colorVar("hl-yellow"),
+        "yellow-dark": colorVar("hl-yellow-dark"),
       },
     },
     fontFamily: fontCatalog.families(),
@@ -100,8 +91,17 @@ var config = {
     },
     extend: {
       boxShadow: {
-        "sidebar-l": "5px 0 10px -5px rgba(0, 0, 0, 0.1)",
-        "panel-t": "0 4px 6px 6px rgba(0, 0, 0, 0.1)",
+        "sidebar-l": "5px 0 10px -5px var(--default-shadow)",
+        "panel-t": "0 4px 6px 6px var(--default-shadow)",
+      },
+      boxShadowColor: {
+        DEFAULT: "var(--default-shadow)",
+        sm: "var(--default-shadow)",
+        md: "var(--default-shadow)",
+        lg: "var(--default-shadow)",
+        xl: "var(--default-shadow)",
+        "2xl": "var(--default-shadow)",
+        inner: "var(--default-shadow)",
       },
       contrast: {
         "105": "1.05",
@@ -143,7 +143,6 @@ var config = {
   },
   plugins: [
     require("./ui/plugins/tw-interactions"),
-    require("./ui/plugins/tw-forms"),
   ],
 }